% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World Cup attendance data.}
\format{
a data frame with 22 observations
\describe{
  \item{Year}{The year of past world cups.}
  \item{Hosts}{Which country hosted the world cup.}
  \item{Totalattendance}{Total attendance at world cup.}
  \item{Matches}{Number of matches.}
  \item{Averageattendance}{Average attendance at world cup.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
World Cup attendance data.
}
\keyword{datasets}
