% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Population data for various countries from 1950 to 2020.}
\format{
a data frame with 19880 observations
\describe{
  \item{Country_Name}{The name of the country.}
  \item{Year}{Years 1950 to 2020.}
  \item{Population}{Population of country in specified year.}
}
}
\source{
\xlsx{World_Populations.xlsx}
}
\usage{
WorldPopulation
}
\description{
Population data for various countries from 1950 to 2020.
}
\keyword{datasets}
